requirements.txt
amqp==2.5.1
atomicwrites==1.3.0
attrs==19.1.0
bcrypt==3.1.4
billiard==3.6.1.0
celery==4.3.0
cffi==1.12.3
Click==7.0
Flask==1.0.2
Flask-SocketIO==3.3.2
graphviz==0.11.1
gunicorn==19.9.0
importlib-metadata==0.19
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.6.4
MarkupSafe==1.1.1
mem-top==0.1.6
more-itertools==7.0.0
newrelic==5.0.2.126
objgraph==3.4.1
pluggy==0.9.0
py==1.8.0
pycparser==2.19
pymongo==3.8.0
pytest==4.4.1
python-chess==0.27.3
python-engineio==3.5.1
python-socketio==3.1.2
pytz==2019.2
redis==3.3.7
six==1.12.0
vine==1.3.0
Werkzeug==0.15.2
zipp==0.5.2
